refactor(crane-mcp): split index.ts into registry + dispatch modules (#843)
`packages/crane-mcp/src/index.ts` was 1115 lines, dominated by a 740-line
ListTools handler (static tool definitions) and a 250-line CallTool
switch (uniform dispatch boilerplate). Decomposed at cohesion boundaries:

- `registry/<category>.ts` — MCP tool definitions co-located with their
  schema + executor wiring via a tiny `makeEntry` factory. Split by
  domain (session, docs, notes, fleet, notifications, skills, memory,
  verify); largest is notes.ts at 183 LOC.
- `tool-runtime.ts` — `ToolEntry` type, `makeEntry` factory,
  `buildRegistry` indexer, `dispatchTool` (replaces the switch with a
  Map lookup; preserves the existing logsTokens-on-success +
  always-log-on-error pattern verbatim).
- `lib/tool-tokens.ts` — `logToolTokens` helper extracted from index.ts.
- `index.ts` — slimmed to 61 lines: server setup, two SDK request
  handlers, main(). All behavior preserved; tool manifest snapshot still
  emits the same 27 tools.

Brings every file in the package under @venturecrane/eslint-config v0.1
caps (≤500 LOC, ≤75 LOC/fn, complexity ≤15, depth ≤4, params ≤5).
`npm run verify` green.
